% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_stack.R
\name{atlas_stack}
\alias{atlas_stack}
\title{get atlas SpatRaster stack}
\usage{
atlas_stack(resolution = 1)
}
\arguments{
\item{resolution}{Numeric: resolution 1 = 5m, 2 = 10m, 4 = 20m}
}
\value{
a SpatRatser stack
}
\description{
get SpatRaster stack with environmental layers (landuse, distance to settlements, distance to kettleholes, distance to rivers, distance to streets)
}
\examples{
\dontrun{
d6atlas::atlas_stack(resolution = 1)
d6atlas::atlas_stack(resolution = 4)
 }
}
