% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bushland_07}
\alias{bushland_07}
\title{bushland}
\format{
A data frame with 130 rows and 6 variables:
\describe{
\item{\code{fid}}{double ID}
\item{\code{BIOTYP8}}{character biotope type}
\item{\code{BIOTYP8_T}}{character biotope text (german)}
\item{\code{Shape_Leng}}{double length in m}
\item{\code{Shape_Area}}{double area in m²}
\item{\code{geometry}}{list geometry column}
}
}
\usage{
bushland_07
}
\description{
sf object containing areas of bushland (biotope type 07)
}
\details{
https://metaver.de/trefferanzeige?docuuid=B57B9F35-AFFF-49F2-BA32-618D1A1CD412
}
\keyword{datasets}
