% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_human_settlements}
\alias{dist_human_settlements}
\title{dist_human_settlements}
\format{
values distance to settlements
\describe{
\item{\code{dist_human_settlements}}{vector with distance values in m}
}
}
\usage{
dist_human_settlements
}
\description{
a vector with spatRatser values
}
\details{
DETAILS
}
\keyword{datasets}
