% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_kettleholes}
\alias{dist_kettleholes}
\title{dist_kettleholes}
\format{
values distance to kettleholes
\describe{
\item{\code{dist_kettleholes}}{vector with distance values in m}
}
}
\usage{
dist_kettleholes
}
\description{
vector with spatRatser values
}
\details{
DETAILS
}
\keyword{datasets}
