% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_rivers}
\alias{dist_rivers}
\title{dist_rivers}
\format{
values distance to rivers
\describe{
\item{\code{dist_rivers}}{vector with distance values in m}
}
}
\usage{
dist_rivers
}
\description{
vector with spatRatser values
}
\details{
DETAILS
}
\keyword{datasets}
