% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_streets}
\alias{dist_streets}
\title{dist_streets}
\format{
values distance to streets
\describe{
\item{\code{dist_streets}}{vector with distance values in m}
}
}
\usage{
dist_streets
}
\description{
vector with spatRatser values
}
\details{
DETAILS
}
\keyword{datasets}
