% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist_water}
\alias{dist_water}
\title{dist_water}
\format{
values distance to waterbodies
\describe{
\item{\code{dist_water}}{vector with distance values in m}
}
}
\usage{
dist_water
}
\description{
vector with spatRatser values
}
\details{
DETAILS
}
\keyword{datasets}
