% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stack.R
\name{extract_stack}
\alias{extract_stack}
\title{get a dataframe with the extracted values from the atlas_stack}
\usage{
extract_stack(sf_point)
}
\arguments{
\item{sf_point}{a sf point with extent of atlas_stack.}
}
\value{
a af object with extracted values
}
\description{
add extracted values from atlas_stack (landuse, distance to settlements, distance to kettleholes, distance to rivers, distance to streets, distance to waterbodies) to the starling_2547 dataframe or to another one
}
\examples{
\dontrun{
d6atlas::extract_stack(sf_point = sf_point)
 }
}
