% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landuse}
\alias{landuse}
\title{landuse}
\format{
values of landuse categories
\describe{
\item{\code{landuse}}{vector with landuse categories}
}
}
\usage{
landuse
}
\description{
a vector with spatRatser values
}
\details{
DETAILS
}
\keyword{datasets}
