% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landuse_reclass_table}
\alias{landuse_reclass_table}
\title{landuse_reclass_table}
\format{
dataframe with five columns
\describe{
\item{\code{biotoptype}}{character biotope main type}
\item{\code{Name_Klasse}}{character biotope main text (german)}
\item{\code{BIOTYP8}}{character biotope sub type}
\item{\code{Name_8st}}{character biotope sub text (german)}
\item{\code{name_class_eng}}{character biotope text (german)}
}
}
\usage{
landuse_reclass_table
}
\description{
dataframe with biotope types with original and english names for classes
}
\details{
DETAILS
}
\keyword{datasets}
