% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ras_list}
\alias{ras_list}
\title{spatRatser extent and crs}
\format{
A list with extent object and crs string
\describe{
\item{\code{extent}}{terra extent object}
\item{\code{crs}}{character crs string}
}
}
\usage{
ras_list
}
\description{
list with extent object and crs string
}
\details{
DETAILS
}
\keyword{datasets}
