% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{stations}
\format{
a dataframe with 16 rows and 19 columns
\describe{
\item{\code{Station_id}}{character Station id}
\item{\code{x_2019}}{numeric x coordinate}
\item{\code{y_2019}}{numeric y coordinate}
\item{\code{z_2019}}{numerix z coordinate}
\item{\code{x_2020}}{numeric x coordinate}
\item{\code{y_2020}}{numeric y coordinate}
\item{\code{z_2020}}{numerix z coordinate}
\item{\code{x_2021}}{numeric x coordinate}
\item{\code{y_2021}}{numeric y coordinate}
\item{\code{z_2021}}{numerix z coordinate}
\item{\code{x_2022}}{numeric x coordinate}
\item{\code{y_2022}}{numeric y coordinate}
\item{\code{z_2022}}{numerix z coordinate}
\item{\code{x_2023}}{numeric x coordinate}
\item{\code{y_2023}}{numeric y coordinate}
\item{\code{z_2023}}{numerix z coordinate}
\item{\code{x_2024}}{numeric x coordinate}
\item{\code{y_2024}}{numeric y coordinate}
\item{\code{z_2024}}{numerix z coordinate}
}
}
\usage{
stations
}
\description{
table with station locations from 2019 to 2024
}
\details{
DETAILS
}
\keyword{datasets}
