% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations_data}
\alias{stations_data}
\title{get stations data frame or sf object}
\usage{
stations_data(year = NULL)
}
\arguments{
\item{year}{numeric filter data by year (2019-2024).}
}
\value{
a data frame or sf object
}
\description{
get the stations dataset with coordinates for each year. If argument year is set, the data will be filtered and automatically transformed to an sf object.
}
\examples{
\dontrun{
d6atlas::stations_data(year = NULL)
 }
}
