% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swamp_04}
\alias{swamp_04}
\title{swamp & wetland}
\format{
A data frame with 35 rows and 6 variables:
\describe{
\item{\code{fid}}{double ID}
\item{\code{BIOTYP8}}{character biotope type}
\item{\code{BIOTYP8_T}}{character biotope text (german)}
\item{\code{Shape_Leng}}{double length in m}
\item{\code{Shape_Area}}{double area in m²}
\item{\code{geometry}}{list geometry column}
}
}
\usage{
swamp_04
}
\description{
sf object containing areas of swamps & wetland (biotope type 04)
}
\details{
https://metaver.de/trefferanzeige?docuuid=B57B9F35-AFFF-49F2-BA32-618D1A1CD412
}
\keyword{datasets}
