% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_02}
\alias{water_02}
\title{Still water courses}
\format{
A data frame with 75 rows and 6 variables:
\describe{
\item{\code{fid}}{double ID}
\item{\code{BIOTYP8}}{character biotope type}
\item{\code{BIOTYP8_T}}{character biotope text (german)}
\item{\code{Shape_Leng}}{double length in m}
\item{\code{Shape_Area}}{double area in m²}
\item{\code{geometry}}{list geometry column}
}
}
\usage{
water_02
}
\description{
sf object containing still water courses (biotope type 02)
}
\details{
https://metaver.de/trefferanzeige?docuuid=B57B9F35-AFFF-49F2-BA32-618D1A1CD412
}
\keyword{datasets}
