% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasauto.R
\name{atlasauto}
\alias{atlasauto}
\title{Transform your data & plot with suitable bounding box}
\usage{
atlasauto(
  data,
  size = 1,
  color = "red",
  scalebar = FALSE,
  north_arrow = FALSE,
  color_intensity = 1
)
}
\arguments{
\item{data}{your data frame, can be filtered or unfiltered}

\item{size}{size of the dots, Default: 1}

\item{color}{color of the dots, Default: "red"}

\item{scalebar}{Boolean: add a scale bar, Default: FALSE}

\item{north_arrow}{Boolean: add a north arrow, Default: FALSE}

\item{color_intensity}{Numeric: control the desaturation of
the color intensity, ranging from 0 (fully
desaturated) to 1 (fully saturated), Default: 1}
}
\value{
a ggplot object with the bounding box of your data
}
\description{
atlasauto aims to provide ATLAS users in Dedelow a fast solution to plot your data after receiving / downloading it. The function automatically transforms X and Y into a sf-object, transforms the coordinates into ETRS89 / UTM zone 33N and crops the underlying map for your needs.
}
\seealso{
\code{\link[sf]{st_as_sf}},\code{\link[sf]{st_transform}}
\code{\link[ggplot2]{CoordSf}}
}
