% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasplot.R
\name{atlasplot}
\alias{atlasplot}
\title{FUNCTION_TITLE}
\usage{
atlasplot(
  color_intensity = 1,
  legend = "bottom",
  legend_x = NULL,
  legend_y = NULL,
  print = FALSE
)
}
\arguments{
\item{color_intensity}{PARAM_DESCRIPTION, Default: 1}

\item{legend}{PARAM_DESCRIPTION, Default: 'bottom'}

\item{legend_x}{PARAM_DESCRIPTION, Default: NULL}

\item{legend_y}{PARAM_DESCRIPTION, Default: NULL}

\item{print}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{between}}
\code{\link[scales]{alpha}}
\code{\link[grDevices]{colorRamp}}
\code{\link[colorspace]{desaturate}},\code{\link[colorspace]{lighten}}
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{CoordSf}}
}
