% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasplot.R
\name{atlasplot}
\alias{atlasplot}
\title{Plot template map Uckermark (ATLAS experimental area)}
\usage{
atlasplot(
  color_intensity = 1,
  scalebar = FALSE,
  north_arrow = FALSE,
  insert_caption = FALSE,
  legend = "bottom",
  legend_x = NULL,
  legend_y = NULL,
  print = FALSE
)
}
\arguments{
\item{color_intensity}{Numeric: control the desaturation of
the color intensity, ranging from 0 (fully
desaturated) to 1 (fully saturated), Default: 1}

\item{scalebar}{Boolean: add a scale bar, Default: FALSE}

\item{north_arrow}{Boolean: add a north arrow, Default: FALSE}

\item{insert_caption}{Boolean: add a caption, Default: FALSE}

\item{legend}{string: Either "bottom", "top", or "none". Otherwise specify
positions via legend_x and legend_y, Default: 'bottom'}

\item{legend_x}{Numeric: Horizontal position of the legend (0 – 1), Default: NULL}

\item{legend_y}{Numeric: Vertical position of the legend (0 – 1), Default: NULL}

\item{print}{Boolean: print the map in the viewer pane, Default: FALSE}
}
\value{
A ggplot object containing a template map for the ATLAS-region
}
\description{
plots a template map for the experimental area of the ATLAS system in the Uckermark, Brandenburg
}
\examples{
\dontrun{
d6atlas::atlasplot()
d6atlas::atlasplot(north_arrow = TRUE, scalebar = TRUE)
 }
}
\seealso{
\code{\link[dplyr]{between}}
\code{\link[scales]{alpha}}
\code{\link[grDevices]{colorRamp}}
\code{\link[colorspace]{desaturate}},\code{\link[colorspace]{lighten}}
\code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{CoordSf}}
}
