% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landuse_reclass_table}
\alias{landuse_reclass_table}
\title{landuse_reclass_table}
\format{
a dataframe with 7755 rows and 5 columns
\describe{
\item{\code{biotoptype }}{numeric biotoptype}
\item{\code{Name_Klasse}}{name of biotoptype in german}
\item{\code{BIOTYP8}}{8 character biotoptype }
\item{\code{Name_8st}}{sub categories in german}
\item{\code{name_class_eng}}{reclass names in english}
}
}
\usage{
landuse_reclass_table
}
\description{
table with landuse categories and a reclass column
}
\details{
DETAILS
}
\keyword{datasets}
